Update Level_1
Added PlayerBehaviour.cs, PlayerCollision.cs, CameraMovement.cs, and a script folder.
Added the geometry dash sprite on the player. Changed the name of the folder "icon" to "sprite".
Limited the player to 1 jump.

Ongoing :
Currently trying to make the sprite rotate while it is in the air (kinda works but pretty weird). To be changed for a better solution.